Externalize non-vscode dependencies
