+++ package.json
@@ -212,6 +212,7 @@
     "@xmldom/xmldom": "^0.9.8",
     "tiktoken": "^1.0.20",
     "undici": "^6.21.2",
+    "vscode-uri": "^3.1.0",
     "xml2js": "^0.6.2"
   }
 }
