+++ src/commands.ts
@@ -1,6 +1,6 @@
 import * as vscode from "vscode";
 import { Disposable } from "./provider";
-import { Profile } from "./profiles";
+import { Profile } from "./contracts";
 import { commands, getConfig, globalStates, internalCommands, ODataMode } from "./configuration";
 import { combineODataUrl } from "./formatting";
 import { QueryRunner } from "./services/QueryRunner";
