+++ src/contracts.ts
@@ -1,3 +1,49 @@
+import { URI } from "vscode-uri";
+
 export interface IFileReader {
     readFile(path: string): Promise<Uint8Array>;
+}
+
+export enum AuthKind {
+    None = "none",
+    Basic = "basic",
+    Bearer = "bearer",
+    ClientCert = "cliencert",
+}
+
+export interface IProfileAuthentication {
+    kind: AuthKind;
+    username?: string;
+    password?: string;
+    token?: string;
+    cert?: URI;
+    key?: URI;
+    pfx?: URI;
+    passphrase?: string;
+}
+
+export interface Profile {
+    name: string;
+    baseUrl: string;
+    auth: IProfileAuthentication;
+    metadata?: string;
+    headers: { [key: string]: string };
+}
+
+export type ODataFormat = "json" | "xml";
+
+export interface IODataMetadataConfigurationMapEntry {
+    url: string;
+    path: string;
+}
+
+export interface IODataMetadataConfiguration {
+    filterNs: string[];
+    removeAnnotations: boolean;
+}
+
+export interface IODataConfiguration {
+    metadata: IODataMetadataConfiguration;
+    defaultFormat: ODataFormat;
+    strictParser: boolean; // Added strictParser setting
 }
